+++ src/Form/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Enforce `max_elements` limits consistently in `Collection::setValue()` and `Collection::submitValue()`
 - Rewind uploaded file stream after reading magic bytes in `FileFormatValidator::validate()` to prevent downstream code from missing the beginning of the stream
 
 ## [3.0.0] - 2026-02-19